+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased]
 
+### Added
+
+- Type declarations for every config. Each one is typed as a Prettier `Config`.
+
 ## [1.0.0] - 2026-09-20
 
 ### Added